resolves #5: removed two unnecessary array typecasting operations

+++ nd2reader/model/__init__.py
@@ -86,8 +86,8 @@ class Image(object):
     @property
     def data(self):
         if self._data is None:
-            # The data is just a flat, 1-dimensional array. We convert it to a 2D array and cast the data points as 16-bit integers
-            self._data = np.reshape(self._raw_data, (self._height, self._width)).astype(np.int64).astype(np.uint16)
+            # The data is just a flat, 1-dimensional array. We convert it to a 2D image here.
+            self._data = np.reshape(self._raw_data, (self._height, self._width))
         return self._data
 
     @property
